eventmanager added
