+++ EmotionCalculator/EmotionCalculator.csproj
@@ -156,6 +156,7 @@
     <Compile Include="EmotionCalculator\Logic\Data\MonthEmotionsIO.cs" />
     <Compile Include="EmotionCalculator\Logic\Data\Songs\RadioPack.cs" />
     <Compile Include="EmotionCalculator\Logic\Data\Songs\SongPack.cs" />
+    <Compile Include="EmotionCalculator\Logic\Events\EventManager.cs" />
     <Compile Include="EmotionCalculator\Logic\MonthManager.cs" />
     <Compile Include="EmotionCalculator\Logic\User\ReadOnlyUserData.cs" />
     <Compile Include="EmotionCalculator\Logic\User\UserLoader.cs" />
